push normalization down to repo
